server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
anthropic>=0.40
pydantic-settings>=2.6
httpx>=0.28
apscheduler>=3.10
pyyaml>=6.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
pytest>=8.0
pytest-asyncio>=0.24
